+++ source/MIBRR/src/RcppExports.cpp
@@ -46,16 +46,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// drawMVN
-Eigen::MatrixXd drawMVN(int n, Eigen::VectorXd meanVec, Eigen::MatrixXd covMat);
-RcppExport SEXP _MIBRR_drawMVN(SEXP nSEXP, SEXP meanVecSEXP, SEXP covMatSEXP) {
+// drawMvn
+Eigen::MatrixXd drawMvn(int n, Eigen::VectorXd meanVec, Eigen::MatrixXd covMat);
+RcppExport SEXP _MIBRR_drawMvn(SEXP nSEXP, SEXP meanVecSEXP, SEXP covMatSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< Eigen::VectorXd >::type meanVec(meanVecSEXP);
     Rcpp::traits::input_parameter< Eigen::MatrixXd >::type covMat(covMatSEXP);
-    rcpp_result_gen = Rcpp::wrap(drawMVN(n, meanVec, covMat));
+    rcpp_result_gen = Rcpp::wrap(drawMvn(n, meanVec, covMat));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -82,6 +82,20 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type mu(muSEXP);
     Rcpp::traits::input_parameter< double >::type lambda(lambdaSEXP);
     rcpp_result_gen = Rcpp::wrap(drawInvGauss(n, mu, lambda));
+    return rcpp_result_gen;
+END_RCPP
+}
+// drawGig
+Eigen::VectorXd drawGig(int n, double lambda, double chi, double psi);
+RcppExport SEXP _MIBRR_drawGig(SEXP nSEXP, SEXP lambdaSEXP, SEXP chiSEXP, SEXP psiSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< double >::type lambda(lambdaSEXP);
+    Rcpp::traits::input_parameter< double >::type chi(chiSEXP);
+    Rcpp::traits::input_parameter< double >::type psi(psiSEXP);
+    rcpp_result_gen = Rcpp::wrap(drawGig(n, lambda, chi, psi));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -121,9 +135,10 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_MIBRR_runGibbs", (DL_FUNC) &_MIBRR_runGibbs, 17},
     {"_MIBRR_drawInvGamma", (DL_FUNC) &_MIBRR_drawInvGamma, 3},
-    {"_MIBRR_drawMVN", (DL_FUNC) &_MIBRR_drawMVN, 3},
+    {"_MIBRR_drawMvn", (DL_FUNC) &_MIBRR_drawMvn, 3},
     {"_MIBRR_calcIncGamma", (DL_FUNC) &_MIBRR_calcIncGamma, 3},
     {"_MIBRR_drawInvGauss", (DL_FUNC) &_MIBRR_drawInvGauss, 3},
+    {"_MIBRR_drawGig", (DL_FUNC) &_MIBRR_drawGig, 4},
     {"_MIBRR_printObsIndices", (DL_FUNC) &_MIBRR_printObsIndices, 6},
     {"_MIBRR_printMissIndices", (DL_FUNC) &_MIBRR_printMissIndices, 6},
     {NULL, NULL, 0}